Refactor to expose functionality as SyncedAccessPoint (#62)
As discussed in #55, I think this refactor exposes the sync functionality in a more "CDK" appropriate way. It removes the capability to sync data to an arbitrary AccessPoint, but I think that makes sense as discussed in the issue above - most importantly because the sync is potentially destructive to any existing data so users should not attempt to use an AccessPoint that has existing content.

I haven't yet updated the README.md but you can see how to use it in integ.default.ts.

Let me know what you think. If you agree, the minor version number should probably be bumped from 0.1 to 0.2 as this changeset would break backward compatibility.

Fixes #55